add: account controller test(partial)

+++ server/src/controllers/account.controller.ts
@@ -1,60 +1,44 @@
-import { AppDataSource } from '../database/data-source'
-import { Account } from '../database/entity'
+import { AppDataSource } from '../database/dataSource'
+import { Account } from '../database/entities'
 import { middleware } from '../middlewares/middleware'
+import HttpException, {
+  ForbiddenStatus,
+  UnauthorizedStatus,
+} from '../utils/exceptions/HttpException'
 
 const accountRepository = AppDataSource.getRepository(Account)
 
 export const createAccount: middleware = async (req, res, next) => {
-  try {
-    const account = accountRepository.create(req.body)
-    const savedAccount = await accountRepository.save(account)
-    return res.send(savedAccount)
-  } catch (err) {
-    console.error(err)
-  }
+  const account = accountRepository.create(req.body)
+  const savedAccount = await accountRepository.save(account)
+  return res.send(savedAccount)
 }
 
 export const updateAccount: middleware = async (req, res, next) => {
-  try {
-    const accountId = req.params.accountId as unknown as number
-    await accountRepository.update(accountId, req.body)
-  } catch (err) {
-    console.error(err)
+  if (!req.decodedToken) {
+    throw new HttpException(ForbiddenStatus)
   }
+  const accountId = req.params.accountId as unknown as number
+  await accountRepository.update(accountId, req.body)
 }
 
 export const deleteAccount: middleware = async (req, res, next) => {
-  try {
-    const accountId = req.params.accountId as unknown as number
-    await accountRepository.delete(accountId)
-  } catch (err) {
-    console.error(err)
+  if (!req.decodedToken) {
+    throw new HttpException(ForbiddenStatus)
   }
+  const accountId = req.params.accountId as unknown as number
+  await accountRepository.delete(accountId)
 }
 
 export const getAccount: middleware = async (req, res, next) => {
-  try {
-    const accountId = req.params.accountId as unknown as number
-    const account = await accountRepository.findOneBy({ id: accountId })
-    return res.send(account)
-  } catch (err) {
-    console.error(err)
+  if (!req.decodedToken) {
+    throw new HttpException(ForbiddenStatus)
   }
-}
+  const accountId = req.params.accountId as unknown as number
 
-/** TODO: create access token and refresh token */
-export const createToken: middleware = async (req, res, next) => {
-  try {
-    const { email, password } = req.body
-  } catch (err) {
-    console.error(err)
+  if (accountId != req.decodedToken.accountId) {
+    throw new HttpException(UnauthorizedStatus)
   }
-}
-
-/** This function is not planned yet */
-export const refreshToken: middleware = async (req, res, next) => {
-  try {
-  } catch (err) {
-    console.error(err)
-  }
+  const account = await accountRepository.findOneBy({ id: accountId })
+  return res.send(account)
 }
